NLS_Fit

Documentation for NLS_Fit.

NLS_Fit.Abstract_Map

NLS_Fit.Abstract_Model2Fit

NLS_Fit.Gaussian_Peak

NLS_Fit.Map_Affine

NLS_Fit.Map_Affine_Monotonic

NLS_Fit.Map_Translate

NLS_Fit.Model2Fit_Empty

NLS_Fit.Model2Fit_Mapped_Parameters

NLS_Fit.Model2Fit_Recalibration

NLS_Fit.Model2Fit_Shared_Parameters

NLS_Fit.Model2Fit_Stacked

NLS_Fit.Model2Fit_TaggedModel

NLS_Fit.accumulate_y!

NLS_Fit.eval_calibrated_x

NLS_Fit.eval_map

NLS_Fit.eval_y

NLS_Fit.get_model

NLS_Fit.get_model

NLS_Fit.get_model_θ

NLS_Fit.get_model_θ

NLS_Fit.get_tagged_data

NLS_Fit.get_tagged_data_type

NLS_Fit.get_tagged_model

NLS_Fit.get_tagged_model_type

NLS_Fit.insert_some_elements

NLS_Fit.parameter_size

NLS_Fit.parameter_size

NLS_Fit.solve_linear_parameters

NLS_Fit.visit

NLS_Fit.Abstract_Map — Type

Map base type.

This type is an abstraction of a $f:X ↦ Y=f(X)$.

Such transformations can be used in various contexts:

calibration task: we have a reference $\hat{X}$ to a calibrated $X$

parameter transformation: we map a parameter vector $\hat{θ}$ to a new one $θ$.

Interface

parameter_size

eval_map

NLS_Fit.Abstract_Model2Fit — Type

Abstract type, base of all model to fit.

Interface

parameter_size

accumulate_y!

NLS_Fit.Gaussian_Peak — Type

Gaussian peak

\[p(x; θ = [h,μ,σ]) = h e^{-\frac{1}{2} \left(\frac{x-μ}{σ} \right)^2}\]

NLS_Fit.Map_Affine — Type

Map_Affine(X_hat_A => X_A, X_hat_B => X_B)

Map_Affine(X_hat_A, X_hat_B)

The second constructor is a shortcut for

Map_Affine(X_hat_A => X_hat_A, X_hat_B => X_hat_B)

Defines an affine transformation:

\[X(\hat{X}) = L_A(\hat{X}) X_A θ_A + L_B(\hat{X}) X_B θ_B\]

where $L_A$ and $L_B$ are the Lagrange basis.

In peculiar, for $(θ_A,θ_B) = (1,1)$ we have an affine map such that $X(\hat{X}_A) = X_A$ and $X(\hat{X}_B) = X_B$.

NLS_Fit.Map_Affine_Monotonic — Type

Map_Affine_Monotonic(X_hat_A => X_A, X_hat_B => X_B)

Map_Affine_Monotonic(X_hat_A, X_hat_B)

The second constructor is a shortcut for

Map_Affine_Monotonic(X_hat_A => X_hat_A, X_hat_B => X_hat_B)

Same as Map_Affine but uses another parametrization that allows to insure monotonic map using simple bound constraints.

The Map_Affine map is defined as follows:

\[X(\hat{X}) = L_A(\hat{X}) X_A θ_A + L_B(\hat{X}) X_B θ_B\]

With this parametrization you have to add the $X_B θ_B ≥ X_A θ_A$ to insure that the map is increasing.

On the other hand, with the Map_Affine_Monotonic parametrization:

\[X(\hat{X}) = L_A(\hat{X}) X_A θ_A + L_B(\hat{X}) ( (X_B-X_A) θ_B + X_A θ_A )\]

where

\[\frac{dX}{d\hat{X}} = \frac{X_B-X_A}{\hat{X}_B-\hat{X}_A} θ_B\]

a simply bound constraint $θ_B ≥ 0$ is sufficient to preserve the increasing or decreasing character. This preserved slope is given by the sign of $\frac{X_B-X_A}{\hat{X}_B-\hat{X}_A}$ which is constant once the structure has been initialized.

In peculiar, for $(θ_A,θ_B) = (1,1)$ we have an affine map such that $X(\hat{X}_A) = X_A$ and $X(\hat{X}_B) = X_B$. Whereas for $(θ_A,θ_B) = (1,0)$ we have a constant one.

NLS_Fit.Map_Translate — Type

Map_Translate(;scale = 1)

Defines a translate transformation:

\[X(\hat{X}) = \hat{X} + scale (θ - 1)\]

In peculiar, for $(θ) = (1)$ the transformation is identity. This is due to the presence of the $θ- 1$ factor. This factor is introduced to stay consistent with other transformations where we try to have identity transform for parameter vector of ones.

NLS_Fit.Model2Fit_Empty — Type

An empty model without parameters. This is useful to initialize summation:

model = Model2Fit_Empty()

for i in 1:n
    model = model + GaussianPeak()
end

NLS_Fit.Model2Fit_Mapped_Parameters — Type

Create a new model, where some parameters are computed using a Abstract_Map.

TODO: not clear how to use it: certainly need a refactoring...

Also see

Model2Fit_Shared_Parameters

NLS_Fit.Model2Fit_Recalibration — Type

Create a calibrable model

Constructors

calibrable_model = Model2Fit_Recalibration(model_to_calibrate, calibration_map)

where model_to_calibrate is an Abstract_Model2Fit and calibration_map is an Abstract_Map

Explanation

The call

eval_y(calibrable_model,X_hat,[θ, θ_map])

is equivalent to

eval_y(model_to_calibrate,eval_calibrated_x(X_hat, θ_map), [θ])

NLS_Fit.Model2Fit_Shared_Parameters — Type

Share a set of parameters

Example

Let's assume that we have an initial model model with θ as parameter vector. If we want θ1, θ3, θ5 to share a same value θshared then:

indices_to_share = [1,3,5]
        
model_with_shared_params = Model2Fit_Shared_Parameters(model, indices_to_share)

Do not forget to udpate parameter vector θ, this can be done as follows:

deleteat!(θ, indices_to_share)
push!(θ, θshared)

Also see

Model2Fit_Mapped_Parameters

NLS_Fit.Model2Fit_Stacked — Type

Define a composite model, where each submodel is responsible of a given ROI

| (1) | (2) | ... |––––––|–––––-| ... | models[1] | model[2] | ...

Domain (1) range is defined as 1:ROIsizes[1] Domain (2) range is defined as (ROIsizes[1]+1):ROI_sizes[2] etc...

NLS_Fit.Model2Fit_TaggedModel — Type

Model2Fit_TaggedModel(model,data)

Tag model and embed data

Extra method

get_tagged_data

get_tagged_model

get_tagged_data_type

NLS_Fit.accumulate_y! — Method

accumulate_y!(::Abstract_Model2Fit,Y::AbstractVector,X::AbstractVector,θ::AbstractVector)::AbstractVector

Accumulate model contribution into vector Y.

Also see :

Abstract_Model2Fit

eval_y

NLS_Fit.eval_calibrated_x — Method

eval_calibrated_x(m::Model2Fit_Recalibration,X_hat::AbstractVector,θ::AbstractVector) -> AbstractVector

Compute the calibrated $X$ from the reference domain $\hat{X}$ for the given transformation parameters $θ$.

NLS_Fit.eval_map — Method

eval_map(m::Abstract_Map,X_hat::AbstractVector,θ::AbstractVector) -> X::AbstractVector

Compute $X=X(\hat{X})$.

NLS_Fit.eval_y — Method

eval_y(m::Abstract_Model2Fit,X::AbstractVector,θ::AbstractVector)::AbstractVector

A convenience function that call accumulate_y! using a zero initialized Y vector. This returned vector contains model values.

Also see :

Abstract_Model2Fit

NLS_Fit.get_model — Method

get_model(mp::Model2Fit_Mapped_Parameters) -> Absatrct_Model2Fit

Get back the wrapped model

NLS_Fit.get_model — Method

get_model(mp::Model2Fit_Shared_Parameters) -> Absatrct_Model2Fit

Get back the wrapped model

NLS_Fit.get_model_θ — Method

get_model_θ(mp::Model2Fit_Mapped_Parameters,θ::AbstractVector) -> θ::AbstractVector

Retrieve the parameter vector θ associated to the wrapped model get_model.

NLS_Fit.get_model_θ — Method

get_model_θ(mp::Model2Fit_Shared_Parameters,θ::AbstractVector) -> θ::AbstractVector

Retrieve the parameter vector θ associated to the wrapped model get_model.

NLS_Fit.get_tagged_data — Method

get_tagged_data(m::Model2Fit_TaggedModel{MODEL,DATA)::DATA

Return embedded data

NLS_Fit.get_tagged_data_type — Method

get_tagged_data_type(m::Abstract_Model2Fit)::DataType

Return embedded data type, or Nothing if m is not a Model2Fit_TaggedModel.

NLS_Fit.get_tagged_model — Method

get_tagged_model(m::Model2Fit_TaggedModel{MODEL,DATA)::DATA

Return tagged model

NLS_Fit.get_tagged_model_type — Method

get_tagged_model_type(m::Abstract_Model2Fit)::DataType

Return wrapped model type, or Nothing if m is not a Model2Fit_TaggedModel.

NLS_Fit.insert_some_elements — Function

insert_some_elements(X::AbstractVector{T},
                     indices::AbstractVector{Int},
                     elements::AbstractVector) -> AbstractVector{T}

Insert elements in X at positions indices.

Note: indices must be strictly ordered.

Example: this functions does a work inverse to the deleteat!

julia> X = Any[1:10;]
10-element Vector{Any}:
  1
  2
  3
  4
  5
  6
  7
  8
  9
 10

julia> indices = Int[2, 3, 5, 6, 9, 10];

julia> elements = Float64[2, 3, 5, 6, 9, 10];

julia> deleteat!(X, indices)
4-element Vector{Any}:
 1
 4
 7
 8

julia> insert_some_elements(X,indices,elements)
10-element Vector{Any}:
  1
  2.0
  3.0
  4
  5.0
  6.0
  7
  8
  9.0
 10.0

NLS_Fit.parameter_size — Method

parameter_size(::Abstract_Map) -> Int

Return θ parameter vector length

NLS_Fit.parameter_size — Method

parameter_size(::Abstract_Model2Fit)::Int

Return length of the expected parameter vector θ

Also see :

Abstract_Model2Fit

NLS_Fit.solve_linear_parameters — Method

Find optimal linear parameters

NLS_Fit.visit — Method

visit(mp::Abstract_Model2Fit,X::AbstractVector,θp::AbstractVector,action::Function) -> nothing

Recursively visit models using a depth-first-order.

For each visited model, performian action which is a function of type

visit_default_action(m::Abstract_Model2Fit,x::AbstractVector,θ::AbstractVector)  -> Bool

If the function returns false the depth-first-search is stopped.

Implementation details

The visit functionality requires these methods to be defined for each visited model:

visitsubmodelsize(model)

visitgetsubmodel(model,submodel_idx)

visitgetY(model,submodel_idx,Y,X,θ)

visitgetX(model,submodel_idx,Y,X,θ)

visitgetθ(model,submodel_idx,Y,X,θ)

There is no need to export these methods.

These functions never modifies Y components, in some cases like Model2Fit_Stacked, visit_get_Y returns a view.