API

Index

NLS_Fit.Abstract_Map

NLS_Fit.Abstract_Model2Fit

NLS_Fit.Abstract_Model2Fit_Peak

NLS_Fit.Gaussian_Peak

NLS_Fit.Map_Affine

NLS_Fit.Map_From_VectFunc

NLS_Fit.Map_Quadratic

NLS_Fit.Map_Translate

NLS_Fit.Model2Fit_Const_Parameters

NLS_Fit.Model2Fit_Empty

NLS_Fit.Model2Fit_Mapped_Parameters

NLS_Fit.Model2Fit_Recalibration

NLS_Fit.Model2Fit_Shared_Parameters

NLS_Fit.Model2Fit_Stacked

NLS_Fit.Model2Fit_TaggedModel

NLS_Fit.Model2Fit_Transformed_Parameters

NLS_Fit.NLS_ForwardDiff_From_Model2Fit

NLS_Fit.Transform_Src_Insert_Dest_Map

NLS_Fit.accumulate_y!

NLS_Fit.create_model_transform_src_insert_dest

NLS_Fit.eval_calibrated_x

NLS_Fit.eval_map

NLS_Fit.eval_y

NLS_Fit.get_calibrated_model

NLS_Fit.get_calibrated_model_θ

NLS_Fit.get_calibration_map

NLS_Fit.get_calibration_map_θ

NLS_Fit.get_map_hat_θ_view

NLS_Fit.get_model

NLS_Fit.get_model

NLS_Fit.get_model

NLS_Fit.get_model

NLS_Fit.get_model_hat_θ_view

NLS_Fit.get_model_θ

NLS_Fit.get_model_θ

NLS_Fit.get_model_θ

NLS_Fit.get_model_θ

NLS_Fit.get_tagged_data

NLS_Fit.get_tagged_data_type

NLS_Fit.get_tagged_model

NLS_Fit.get_tagged_model_type

NLS_Fit.insert_some_elements

NLS_Fit.parameter_size

NLS_Fit.parameter_size

NLS_Fit.solve_linear_parameters

NLS_Fit.visit

Doc

Public

NLS_Fit.Abstract_Model2Fit — Type

Abstract type, base of all model to fit.

Interface

The interface is very simple

parameter_size

accumulate_y!

Convenience functions

eval_y

NLS_Fit.Abstract_Model2Fit_Peak — Type

A specialization of Abstract_Model2Fit dedicated to peak types.

TODO: in the future some extra methods are to be defined like

peak_position

peak_area...

NLS_Fit.Gaussian_Peak — Type

Gaussian peak

\[p(x; θ = [h,μ,σ]) = h e^{-\frac{1}{2} \left(\frac{x-μ}{σ} \right)^2}\]

NLS_Fit.Map_Affine — Type

Defines an affine transformation:

\[X(\hat{X}) = L_A(\hat{X}) X_A θ_A + L_B(\hat{X}) X_B θ_B\]

where $L_A$ and $L_B$ are the Lagrange basis.

Constructors

Scaled interpolation

Map_Affine(X_hat_A => X_A, X_hat_B => X_B)

for $(θ_A,θ_B) = (1,1)$ we have an affine map such that $X(\hat{X}_A) = X_A$ and $X(\hat{X}_B) = X_B$.

Usual interpolation

The second constructor

Map_Affine(X_hat_A, X_hat_B)

is a shortcut for

Map_Affine(X_hat_A => 1, X_hat_B => 1)

that defines an usual interpolation (without scaling): $(θ_A,θ_B)$ are simply the values at $(X_A,X_B)$

NLS_Fit.Map_From_VectFunc — Type

An Abstract_Map instance that can be easily created by providing a Julia function

Constructor

Map_From_VectFunc(n_θ,f)

where $(X,θ)->Y=f(X,θ)$ is the wrapped function.

Example

using NLS_Fit

f(X,θ) = θ[1]*X

map = Map_From_VectFunc(1,f)

X = Float64[1,2,3]

θ = Float64[4]

eval_map(map,X,θ)

# output
3-element Vector{Float64}:
  4.0
  8.0
 12.0

NLS_Fit.Map_Quadratic — Type

Map_Quadratic(X_hat_A => X_A, X_hat_B => X_B, X_hat_C => X_C)

Map_Quadratic(X_hat_A, X_hat_B, X_hat_C)

Same semantic as Map_Affine

NLS_Fit.Map_Translate — Type

Map_Translate(;scale = 1)

Defines a translate transformation:

\[X(\hat{X}) = \hat{X} + scale (θ - 1)\]

In peculiar, for $(θ) = (1)$ the transformation is identity. This is due to the presence of the $θ- 1$ factor. This factor is introduced to stay consistent with other transformations where we try to have identity transform for parameter vector of ones.

NLS_Fit.Model2Fit_Const_Parameters — Type

Set some parameters to constant values

Example

Let's assume that we have an initial model model with θ as parameter vector. If we want θ1, θ3, θ5 to be constant, proceed as follows

indices = [1,3,5]
values = Float64[1,2,3]
        
model_with_const_params = Model2Fit_Const_Parameters(model, indices, values)

Do not forget to udpate parameter vector θ, this can be done as follows:

deleteat!(θ, indices)

Also see

NLS_Fit.Model2Fit_Empty — Type

An empty model without parameters. This is useful to initialize summation:

model = Model2Fit_Empty()

for i in 1:n
    model = model + GaussianPeak()
end

NLS_Fit.Model2Fit_Mapped_Parameters — Type

Create a new model, where some parameters are computed using a Abstract_Map.

TODO: not clear how to use it: certainly need a refactoring...

Also see

Model2Fit_Shared_Parameters

NLS_Fit.Model2Fit_Recalibration — Type

Create a calibrable model

More precisely, from an initial model $m$ and a map $f$, create a new model $\hat{m}$ such that:

\[\hat{m}(\hat{X},\hat{\theta}=[\hat{\theta}_m,\hat{\theta}_f]) = m(X=f_{\hat{\theta}_f}(\hat{X}),\hat{\theta}_m)\]

In Julia code this means that

eval_y(calibrable_model,X_hat,[θ̂m, θ̂f])

is replaced by

eval_y(model_to_calibrate, eval_calibrated_x(X_hat, θ̂f), [θ̂m])

Constructors

calibrable_model = Model2Fit_Recalibration(model_to_calibrate, calibration_map)

where model_to_calibrate is an Abstract_Model2Fit and calibration_map is an Abstract_Map

Also see

get_calibration_map

get_calibration_map_θ

get_calibrated_model

get_calibrated_model_θ

eval_calibrated_x

NLS_Fit.Model2Fit_Shared_Parameters — Type

Share a set of parameters

Example

Let's assume that we have an initial model model with θ as parameter vector. If we want θ1, θ3, θ5 to share a same value θshared then:

indices_to_share = [1,3,5]
        
model_with_shared_params = Model2Fit_Shared_Parameters(model, indices_to_share)

Do not forget to udpate parameter vector θ, this can be done as follows:

deleteat!(θ, indices_to_share)
push!(θ, θshared)

Also see

Model2Fit_Mapped_Parameters

NLS_Fit.Model2Fit_Stacked — Type

Define a composite model, where each submodel is responsible of a given ROI

| (1) | (2) | ... |––––––|–––––-| ... | models[1] | model[2] | ...

Domain (1) range is defined as 1:ROIsizes[1] Domain (2) range is defined as (ROIsizes[1]+1):ROI_sizes[2] etc...

NLS_Fit.Model2Fit_TaggedModel — Type

Model2Fit_TaggedModel(model,data)

Tag model and embed data

Extra method

get_tagged_data

get_tagged_model

get_tagged_data_type

NLS_Fit.Model2Fit_Transformed_Parameters — Type

Create a new model, where the parameter vector $\theta$ is computed using an Abstract_Map:

\[ \hat{\theta}\mapsto \theta = f_{\hat{\theta}_f}(\hat{\theta}_m)\]

The resulting model is thus:

\[\hat{m}(X,\hat{\theta}=[\hat{\theta}_m,\hat{\theta}_f]) = m(X,\theta = f_{\hat{\theta}_f}(\hat{\theta}_m))\]

Constructor

General case

Model2Fit_Transformed_Parameters(model::Abstract_Model2Fit,
                                 map_domain_size::Int,
                                 map::Abstract_Map)

map_domain_size is the expected $\hat{\theta}_m$ size. This quantity is used to implement model parameter_size() method. The reason is that the current Abstract_Map interface only provides $\hat{\theta}_f$ size (its parameter_size() method) but tells nothing about $\hat{\theta}_m$ size.

Also, please note that $\theta$ size is not necessary equal to $\hat{\theta}_m$ size

Using Transform_Src_Insert_Dest_Map

Very often this structure is used with a map of type Transform_Src_Insert_Dest_Map. In that case it is simpler to use a dedicated function: create_model_transform_src_insert_dest (follow this link to see an example).

Example

TODO: add an example using the general constructor

Extra methods

Beside the regular Abstract_Model2Fit interface, you can use these extra methods:

get_model(model::Model2Fit_Transformed_Parameters)

get_model_θ(hat_model::Model2Fit_Transformed_Parameters,hat_θ::AbstractVector)

Not exported

These methods are not exported but may be useful:

get_model_hat_θ_view(hat_model::Model2Fit_Transformed_Parameters,hat_θ::AbstractVector)

get_map_hat_θ_view(hat_model::Model2Fit_Transformed_Parameters,hat_θ::AbstractVector)

NLS_Fit.NLS_ForwardDiff_From_Model2Fit — Type

A wrapper that allows to use the NLS_Solver.jl package to solve the nonlinear least squares problem associated to a model plus its (X,Y) data.

You must construct an instance of nls problem as follows:

NLS_ForwardDiff_From_Model2Fit(fit_model::MODEL2FIT_TYPE,
                               X::X_TYPE,
                               Y::Y_TYPE)

You can then use the NLS_Solver package as usual, but with prefixed NLS_Solver.

nls = NLS_ForwardDiff_From_Model2Fit(model,X,Y)
conf = NLS_Solver.LevenbergMarquardt_Conf()
result = NLS_Solver.solve(nls,θ_init,conf)

CAVEAT:

Please note that you should not explicitly

using NLS_Solver

as this package is reexported from NLS_Fit. Doing so avoids potential version compatibility problems.

NLS_Fit.accumulate_y! — Method

accumulate_y!(::Abstract_Model2Fit,Y::AbstractVector,X::AbstractVector,θ::AbstractVector)::AbstractVector

Accumulate model contribution into vector Y.

Also see

Abstract_Model2Fit

eval_y

NLS_Fit.create_model_transform_src_insert_dest — Method

This function creates an Model2Fit_Transformed_Parameters instance from a map of type Transform_Src_Insert_Dest_Map.

function create_model_transform_src_insert_dest(model, map, src=>dest) -> model

Example

using NLS_Fit

# Two Gaussian peaks with σ(μ) an affine function
#
model = Gaussian_Peak() + Gaussian_Peak() 

# model θ is:
# idx:  1   2  3   4   5   6
#  θ : h1, μ1, σ1, h2, μ2, σ2
#
# hence:
#
src  = [2,5]
dest = [3,6]

# The σ(μ) relation: σ(μ=1)=1 and σ(μ=100)=10 for [θ̂A, θ̂B]=[1, 1]
#
f_σ_μ = Map_Affine(1.0=>1.0,100.0=>10.0)

# Now the model_σ_μ parameters are:
#
# [ h1, μ1, h2, μ2, θ̂A, θ̂B ]
#
# the initial model is called with:
#
# [ h1, μ1, σ1 = affine(μ1, θ̂A, θ̂B), h2, μ2, , σ2 = affine(μ2, θ̂A, θ̂B) ]
#
model_σ_μ = create_model_transform_src_insert_dest(model,f_σ_μ,src=>dest)

# Examine that the compute θ takes into account the σ(μ) dependance:
#
θ̂ = Float64[1,5, 2, 95, 1, 1]
θ = get_model_θ(model_σ_μ, θ̂)

# output
6-element Vector{Float64}:
  1.0
  5.0
  1.3636363636363638
  2.0
 95.0
  9.545454545454547

NLS_Fit.eval_calibrated_x — Method

eval_calibrated_x(m::Model2Fit_Recalibration,X_hat::AbstractVector,θ::AbstractVector) -> AbstractVector

Compute the calibrated $X$ from the reference domain $\hat{X}$ for the given transformation parameters $θ$.

See: Model2Fit_Recalibration

NLS_Fit.eval_map — Method

eval_map(m::Abstract_Map,X_hat::AbstractVector,hat_θ::AbstractVector) -> X::AbstractVector

Compute $X=X(\hat{X})$.

See also

Abstract_Map

NLS_Fit.eval_y — Method

eval_y(m::Abstract_Model2Fit,X::AbstractVector,θ::AbstractVector)::AbstractVector

A convenience function that call accumulate_y! using a zero initialized Y vector. This returned vector contains model values.

Also see

Abstract_Model2Fit

NLS_Fit.get_calibrated_model — Method

get_calibrated_model(model::Model2Fit_Recalibration)::Abstract_Model2Fit

Retrieve the underlying model to calibrate

See: Model2Fit_Recalibration

NLS_Fit.get_calibrated_model_θ — Method

get_calibrated_model_θ(model::Model2Fit_Recalibration,
                       θ::AbstractVector)::AbstractVector

Retrieve the parameter vector of the underlying model to calibrate

See: Model2Fit_Recalibration

NLS_Fit.get_calibration_map — Method

get_calibration_map(model::Model2Fit_Recalibration)::Abstract_Map

Retrieve the calibration map, of type Abstract_Map.

See: Model2Fit_Recalibration

NLS_Fit.get_calibration_map_θ — Method

get_calibration_map_θ(model::Model2Fit_Recalibration,
                      θ::AbstractVector)::AbstractVector

Retrieve the calibration map parameters

See: Model2Fit_Recalibration

NLS_Fit.get_model — Method

get_model(model::Model2Fit_Const_Parameters) -> Absatrct_Model2Fit

Get back the wrapped model

NLS_Fit.get_model — Method

get_model(mp::Model2Fit_Mapped_Parameters) -> Absatrct_Model2Fit

Get back the wrapped model

NLS_Fit.get_model — Method

get_model(mp::Model2Fit_Shared_Parameters) -> Absatrct_Model2Fit

Get back the wrapped model

NLS_Fit.get_model — Method

get_model(mp::Model2Fit_Transformed_Parameters)::Abstract_Model2Fit

Get back the wrapped model

See Model2Fit_Transformed_Parameters

NLS_Fit.get_model_θ — Method

get_model_θ(mp::Model2Fit_Const_Parameters,θ::AbstractVector) -> θ::AbstractVector

Retrieve the parameter vector θ associated to the wrapped model get_model.

NLS_Fit.get_model_θ — Method

get_model_θ(mp::Model2Fit_Mapped_Parameters,θ::AbstractVector) -> θ::AbstractVector

Retrieve the parameter vector θ associated to the wrapped model get_model.

NLS_Fit.get_model_θ — Method

get_model_θ(mp::Model2Fit_Shared_Parameters,θ::AbstractVector) -> θ::AbstractVector

Retrieve the parameter vector θ associated to the wrapped model get_model.

NLS_Fit.get_model_θ — Method

get_model_θ(mp::Model2Fit_Transformed_Parameters, hat_θ::AbstractVector) -> θ::AbstractVector

Retrieve the parameter vector θ associated to the wrapped model get_model.

NLS_Fit.get_tagged_data — Method

get_tagged_data(m::Model2Fit_TaggedModel{MODEL,DATA)::DATA

Return embedded data

NLS_Fit.get_tagged_data_type — Method

get_tagged_data_type(m::Abstract_Model2Fit)::DataType

Return embedded data type, or Nothing if m is not a Model2Fit_TaggedModel.

NLS_Fit.get_tagged_model — Method

get_tagged_model(m::Model2Fit_TaggedModel{MODEL,DATA)::DATA

Return tagged model

NLS_Fit.get_tagged_model_type — Method

get_tagged_model_type(m::Abstract_Model2Fit)::DataType

Return wrapped model type, or Nothing if m is not a Model2Fit_TaggedModel.

NLS_Fit.insert_some_elements — Function

insert_some_elements(X::AbstractVector{T},
                     indices::AbstractVector{Int},
                     elements::AbstractVector) -> AbstractVector{T}

Insert elements in X at positions indices.

Note: indices must be strictly ordered.

Example: this functions does a work inverse to the deleteat!

julia> X = Any[1:10;]
10-element Vector{Any}:
  1
  2
  3
  4
  5
  6
  7
  8
  9
 10

julia> indices = Int[2, 3, 5, 6, 9, 10];

julia> elements = Float64[2, 3, 5, 6, 9, 10];

julia> deleteat!(X, indices)
4-element Vector{Any}:
 1
 4
 7
 8

julia> insert_some_elements(X,indices,elements)
10-element Vector{Any}:
  1
  2.0
  3.0
  4
  5.0
  6.0
  7
  8
  9.0
 10.0

NLS_Fit.parameter_size — Method

parameter_size(::Abstract_Model2Fit)::Int

Return length of the expected parameter vector θ

Also see

Abstract_Model2Fit

NLS_Fit.parameter_size — Method

parameter_size(::Abstract_Map) -> Int

Return $\hat{θ}$ parameter vector length

See also

Abstract_Map

NLS_Fit.solve_linear_parameters — Method

Find optimal linear parameters

NLS_Fit.visit — Method

visit(mp::Abstract_Model2Fit,X::AbstractVector,θp::AbstractVector,action::Function) -> nothing

Recursively visit models using a depth-first-order.

For each visited model, performian action which is a function of type

visit_default_action(m::Abstract_Model2Fit,x::AbstractVector,θ::AbstractVector)  -> Bool

If the function returns false the depth-first-search is stopped.

Implementation details

The visit functionality requires these methods to be defined for each visited model:

visit_submodel_size(model)

visit_get_submodel(model,submodel_idx)

visit_get_Y(model,submodel_idx,Y,X,θ)

visit_get_X(model,submodel_idx,Y,X,θ)

visit_get_θ(model,submodel_idx,Y,X,θ)

There is no need to export these methods.

These functions never modifies Y components, in some cases like Model2Fit_Stacked, visit_get_Y returns a view.

Private

NLS_Fit.Abstract_Map — Type

abstract type Abstract_Map end

This type is an abstraction of a

\[f:X ↦ Y=f_{\hat{\theta}}(X)\]

where $\hat{\theta}$ is map parameter vector.

Such transformations can be used in various contexts:

calibration task: we have a reference $\hat{X}$ to a calibrated $X$, see Model2Fit_Recalibration.

parameter transformation: we map a parameter vector $\hat{θ}$ to a new one $θ$, see Model2Fit_Transformed_Parameters.

Interface

parameter_size

eval_map

NLS_Fit.Transform_Src_Insert_Dest_Map — Type

Define a map $g$ of the form

\[\theta = g_{\hat{\theta}_f}(\hat{\theta}_m)\]

Where the resulting $\theta$ vector is computed as follows:

compute $\tau=f_{\hat{\theta}_f}(\hat{\theta}_m[\text{src}])$ an intermediate vector $\tau$ using the wrapped map $f$ action on a subset of the $\hat{\theta}_m$ vector. This subset being defined by an index array src.

create the resulting $\theta$ by inserting components of $\tau$ into the $\hat{\theta}_m$ vector at indices stored in an index array dest (src and dest have the same length).

This is certainly more comprehensible with an example.

Example

using NLS_Fit

f(θ̂m_src,θ̂f) = θ̂f[1] .+ θ̂f[2] * θ̂m_src

f_map = Map_From_VectFunc(2,f)
θ̂f    = Float64[0,-1] # change sign

# insert -θ̂m[4], at position 2
# insert -θ̂m[1], at position 3
src  = [4,1]
dest = [2,3]

# delete elements that are going to be computed & inserted 
#
# we see that elements of `src` indices are kept, 
# whereas elements of `dest` indices are removed
#
θ̂m = Float64[1:5;]       #  [1,2,3,4,5] 
θ̂m = deleteat!(θ̂m,dest)  #  [1,4,5]

map_s_d = NLS_Fit.Transform_Src_Insert_Dest_Map(f_map,src=>dest)

θ = eval_map(map_s_d,θ̂m,θ̂f)

# output
5-element Vector{Float64}:
  1.0
 -4.0
 -1.0
  4.0
  5.0

Constructor

Transform_Src_Insert_Dest_Map(f_map, src=>dest)

f_map is the stored Abstract_Map map,

src=>dest is a pair of two Int vectors used to define indices. Note that we must have src ∩ dest = ∅. The reason is that as dest indices are removed, there would no more associated src component to apply $f$

Also see

This map can be directly used by Model2Fit_Transformed_Parameters.

NLS_Fit.get_map_hat_θ_view — Method

get_model_hat_θ_view(hat_model::Model2Fit_Transformed_Parameters,hat_θ::AbstractVector)

Return a view on $\hat{\theta}_f$ knowing $\hat{\theta}=[\hat{\theta}_m,\hat{\theta}_f]$

NLS_Fit.get_model_hat_θ_view — Method

get_model_hat_θ_view(hat_model::Model2Fit_Transformed_Parameters,hat_θ::AbstractVector)

Return a view on $\hat{\theta}_m$ knowing $\hat{\theta}=[\hat{\theta}_m,\hat{\theta}_f]$